refactor(toolbox): share game-status polling and cache the opened game
Both toolbox tools had the same hand-rolled load + visibilitychange
re-read (users routinely open a tool before launching the game, so the
"game not running" banner has to clear itself). Extract it as
useGameStatus.

On the backend, cache the opened game process instead of rediscovering
it per call: the staleness watchers land in open_game every 5 seconds,
and a fresh discovery walks every process on the system plus every
module of the game just to read a few bytes. The cache is revalidated
with a 4-byte read and rebuilt on failure, so a game restart is still
picked up. The lock is released before any syscall - holding it across
the process-table walk serialized every poller behind the slowest one,
worst exactly when the game is closed and the walk finds nothing.
